Moved Direction to types, added new types
